Make it easier to set up configuration
- Update the README with steps on getting a developer tenant
- Update README with steps on creating sharepoint site
- Reduces specificity of how user inputs data into sharepoint configuration

Previously, the sharepoint configuration item in Innovator needed to be built very specifically. The tenant ID property in particular expected a login URL, but seemed to imply that it only needed the string of the ID itself. I've now added a check to see if the format of the property is a URL or just an ID. If it's just an ID, we append the login URL to the beginning of it. I also added some checks to let the sharepoint_site property accept the URL of the default sharepoint site instead of needing the user to make a new site.